requirements.txt
appnope==0.1.0
awscli==1.15.19
backports.shutil-get-terminal-size==1.0.0
botocore==1.10.19
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.7
coverage==4.5.1
decorator==4.3.0
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
idna==2.6
ipdb==0.11
ipython==5.7.0
ipython-genutils==0.2.0
jmespath==0.9.3
mock==2.0.0
nose2==0.7.4
pathlib2==2.3.2
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
Pygments==2.2.0
python-dateutil==2.7.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
tabulate==0.8.2